backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
matplotlib==3.8.2
seaborn==0.13.0
ta==0.11.0
plotly==5.18.0
requests==2.31.0
sqlalchemy==2.0.23
joblib==1.3.2
